make speech a module
