+++ src/Chat.tsx
@@ -7,8 +7,8 @@ import { Subscription } from 'rxjs/Subscription';
 import { Activity, Media, IBotConnection, User, MediaType, DirectLine, DirectLineOptions, CardActionTypes } from 'botframework-directlinejs';
 import { createStore, ChatActions } from './Store';
 import { Provider } from 'react-redux';
-import { SpeechOptions } from './Speech/SpeechOptions';
-import { SpeechRecognizer, SpeechSynthesizer } from './Speech/SpeechSingleton';
+import { SpeechOptions } from './SpeechOptions';
+import { Speech } from './SpeechModule';
 
 export interface FormatOptions {
     showHeader?: boolean
@@ -86,8 +86,8 @@ export class Chat extends React.Component<ChatProps, {}> {
             this.store.dispatch<ChatActions>({ type: 'Set_Send_Typing', sendTyping: props.sendTyping });
 
         if (props.speechOptions) {
-            SpeechRecognizer.setSpeechRecognizer(props.speechOptions.speechRecognizer);
-            SpeechSynthesizer.setSpeechSynthesizer(props.speechOptions.speechSynthesizer);
+            Speech.SpeechRecognizer.setSpeechRecognizer(props.speechOptions.speechRecognizer);
+            Speech.SpeechSynthesizer.setSpeechSynthesizer(props.speechOptions.speechSynthesizer);
         }
     }
