+++ src/CognitiveServices/lib.ts
@@ -1,0 +1,2 @@
+export { SpeechRecognizer } from './SpeechRecognition'
+export { SpeechSynthesizer } from './SpeechSynthesis'
