--- src/Speech/CognitiveServices/lib.ts
@@ -1,2 +1,0 @@
-export { SpeechRecognizer } from "./SpeechRecognition"
-export { SpeechSynthesizer } from "./SpeechSynthesis"
