+++ src/SpeechOptions.ts
@@ -1,0 +1,6 @@
+import { Speech } from './SpeechModule';
+
+export class SpeechOptions {
+    public speechRecognizer: Speech.ISpeechRecognizer;
+    public speechSynthesizer: Speech.ISpeechSynthesizer;
+}
